requirements.txt
anthropic>=0.45.0
python-telegram-bot>=21.0
httpx>=0.27.0
playwright>=1.40.0
trafilatura>=1.8.0
beautifulsoup4>=4.12.0
dateparser>=1.2.0
rapidfuzz>=3.6.0
tldextract>=5.1.0
pandas>=2.2.0
pyarrow>=15.0.0
openpyxl>=3.1.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
streamlit>=1.35.0
python-jobspy>=1.1.0
python-docx>=1.1.0
reportlab>=4.1.0
